requirements.txt
Django==4.1.4
sqlparse==0.4.3
psycopg2==2.9.5
djangorestframework~=3.14.0